Feat/add game ending

+++ star-dash/star-dash/GameEngine/Systems/Basic/PositionSystem.swift
@@ -42,7 +42,7 @@ class PositionSystem: System {
         }
 
         let angle = atan2(direction.dy, direction.dx)
-        let newRotation = angle - CGFloat(Double.pi / 2)
+        let newRotation = angle - Double.pi / 2
         positionComponent.rotation = newRotation
     }
